feat(#105): Segregate between Scraper and Parser.
